move logging setup out of main
